[WIP] move offroute detector, faster route detector and snap to route classes into the internal package
