+++ libandroid-navigation/src/main/java/com/mapbox/services/android/navigation/v5/internal/navigation/NavigationTelemetry.java
@@ -22,7 +22,7 @@ import com.mapbox.services.android.navigation.v5.internal.navigation.metrics.Nav
 import com.mapbox.services.android.navigation.v5.internal.navigation.metrics.RerouteEvent;
 import com.mapbox.services.android.navigation.v5.internal.navigation.metrics.SessionState;
 import com.mapbox.services.android.navigation.v5.internal.navigation.metrics.TelemetryEvent;
-import com.mapbox.services.android.navigation.v5.internal.navigation.routeprogress.MetricsRouteProgress;
+import com.mapbox.services.android.navigation.v5.internal.routeprogress.MetricsRouteProgress;
 import com.mapbox.services.android.navigation.v5.navigation.MapboxNavigation;
 import com.mapbox.services.android.navigation.v5.navigation.MapboxNavigationOptions;
 import com.mapbox.services.android.navigation.v5.routeprogress.RouteProgress;
