+++ libandroid-navigation/src/main/java/com/mapbox/services/android/navigation/v5/internal/route/FasterRouteDetector.java
@@ -1,4 +1,4 @@
-package com.mapbox.services.android.navigation.v5.route;
+package com.mapbox.services.android.navigation.v5.internal.route;
 
 import android.location.Location;
 
@@ -6,6 +6,7 @@ import com.mapbox.api.directions.v5.models.DirectionsResponse;
 import com.mapbox.api.directions.v5.models.DirectionsRoute;
 import com.mapbox.api.directions.v5.models.LegStep;
 import com.mapbox.api.directions.v5.models.RouteLeg;
+import com.mapbox.services.android.navigation.v5.route.FasterRoute;
 import com.mapbox.services.android.navigation.v5.routeprogress.RouteProgress;
 import com.mapbox.services.android.navigation.v5.routeprogress.RouteStepProgress;
